BAU: Move another method only used in one place
An interim step in a bigger refactor: the point here is to move all the logic in the same place to make it easier to see what can be tidied up
